+++ lib/src/clixon_proto_client.c
@@ -76,7 +76,7 @@
 #include "clixon_stream.h"
 #include "clixon_err_string.h"
 #include "clixon_xml_nsctx.h"
-#include "clixon_xml_map.h"
+#include "clixon_xml_bind.h"
 #include "clixon_xml_sort.h"
 #include "clixon_xml_io.h"
 #include "clixon_netconf_lib.h"